only version is allowed for docker
