+++ component-resolver-core/src/main/java/org/octopusden/octopus/escrow/dto/VersionExpressionContext.java
@@ -1,0 +1,15 @@
+package org.octopusden.octopus.escrow.dto;
+
+public class VersionExpressionContext {
+    private final String version;
+
+    public VersionExpressionContext(String version) {
+        this.version = version;
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+
+}
